+++ shared/SetsnReps.Core/DTOs/Workout/ExerciseSet.cs
@@ -22,6 +22,6 @@ public class ExerciseSet
     /// <summary>
     /// The weight used for the set execution item (if not DurationExecution).
     /// </summary>
-    public float? Weight { get; set; }
+    public decimal? Weight { get; set; }
 
 }
